kafka_transformer/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
pendulum==3.0.0
